requirements.txt
pandas
azure-storage-blob
scikit-learn
